% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hx_page.R
\name{hx_head}
\alias{hx_head}
\title{Specify additional head elements for an htmxr page}
\usage{
hx_head(..., title = "htmxr page")
}
\arguments{
\item{...}{tags to include in the head (stylesheets, scripts, meta, etc.)}

\item{title}{page title}
}
\value{
A list with class \code{hx_head}, to be passed to \code{\link[=hx_page]{hx_page()}}.
}
\description{
Wraps tags to be included in the page head when passed to \code{\link[=hx_page]{hx_page()}}.
}
