% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hx_serve_assets.R
\name{hx_serve_assets}
\alias{hx_serve_assets}
\title{Serve htmxr static assets}
\usage{
hx_serve_assets(api)
}
\arguments{
\item{api}{a plumber2 API object}
}
\value{
the API object (modified, for piping)
}
\description{
Configures a plumber2 API to serve htmxr's static assets
(htmx JavaScript library) at \verb{/htmxr/assets/}.
}
