% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tags}
\alias{p}
\alias{h1}
\alias{h2}
\alias{h3}
\alias{h4}
\alias{h5}
\alias{h6}
\alias{a}
\alias{br}
\alias{div}
\alias{span}
\alias{pre}
\alias{code}
\alias{img}
\alias{strong}
\alias{em}
\alias{hr}
\alias{tag}
\alias{tagList}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{htmltools}{\code{\link[htmltools:a]{a()}}, \code{\link[htmltools:br]{br()}}, \code{\link[htmltools:code]{code()}}, \code{\link[htmltools:div]{div()}}, \code{\link[htmltools:em]{em()}}, \code{\link[htmltools:h1]{h1()}}, \code{\link[htmltools:h2]{h2()}}, \code{\link[htmltools:h3]{h3()}}, \code{\link[htmltools:h4]{h4()}}, \code{\link[htmltools:h5]{h5()}}, \code{\link[htmltools:h6]{h6()}}, \code{\link[htmltools:hr]{hr()}}, \code{\link[htmltools:img]{img()}}, \code{\link[htmltools:p]{p()}}, \code{\link[htmltools:pre]{pre()}}, \code{\link[htmltools:span]{span()}}, \code{\link[htmltools:strong]{strong()}}, \code{\link[htmltools:tag]{tag()}}, \code{\link[htmltools:tagList]{tagList()}}, \code{\link[htmltools:tags]{tags}}}
}}

